+++ src/uci.rs
@@ -104,7 +104,7 @@ fn parse_limits(color: Color, tokens: &[&str]) -> Limits {
             };
 
             match name {
-                "depth" => return Limits::FixedDepth(value as usize),
+                "depth" => return Limits::FixedDepth(value as i32),
                 "movetime" => return Limits::FixedTime(value),
 
                 "wtime" if Color::White == color => main = value,
